+++ unfccc_di_api/unfccc_di_api.py
@@ -22,6 +22,32 @@ import pandas as pd
 import requests
 import treelib
 
+# mapping from gas as simple string to subscript-format used by UNFCCC DI API
+GAS_MAPPING = {
+    "CH4": "CH₄",
+    "CO2": "CO₂",
+    "N2O": "N₂O",
+    "NF3": "NF₃",
+    "SF6": "SF₆",
+    "CF4": "CF₄",
+    "C2F6": "C₂F₆",
+    "c-C3F6": "c-C₃F₆",
+    "C3F8": "C₃F₈",
+    "c-C4F8": "c-C₄F₈",
+    "C4F10": "C₄F₁₀",
+    "C5F12": "C5F₁₂",  # this seems to be a typo in the UNFCCC API
+    "C6F14": "C₆F₁₄",
+    "C10F18": "C₁₀F₁₈",
+    "NH3": "NH₃",
+    "NOx": "NOₓ",
+    "SO2": "SO₂",
+}
+
+# mapping of subscript notation to ASCII string
+NORMALSCRIPT = "0123456789x"
+SUBSCRIPT = "₀₁₂₃₄₅₆₇₈₉ₓ"
+MAKE_ASCII = str.maketrans(SUBSCRIPT, NORMALSCRIPT)
+
 
 class UNFCCCApiReader:
     """Provides simplified unified access to the Flexible Query API of the UNFCCC data
@@ -71,6 +97,7 @@ class UNFCCCApiReader:
         party_code: str,
         gases: typing.Union[typing.List[str], None] = None,
         progress: bool = False,
+        normalize_gas_names: bool = True,
     ) -> pd.DataFrame:
         """Query the UNFCCC for data.
 
@@ -81,9 +108,13 @@ class UNFCCCApiReader:
             :py:attr:`~UNFCCCApiReader.parties`.
         gases : list of str, optional
             Limit the query to these gases. For possible values, see
-            :py:attr:`~UNFCCCApiReader.gases`. Default: query for all gases.
+            :py:attr:`~UNFCCCApiReader.gases`. Accepts subscripts ("N₂O")
+            as well as ASCII-strings ("N2O"). Default: query for all gases.
         progress : bool
             Display a progress bar. Requires the :py:mod:`tqdm` library. Default: false.
+        normalize_gas_names : bool, optional
+            If :obj:`True`, return gases as ASCII strings ("N2O").
+            Else, return native UNFCCC notation ("N₂O"). Default: true.
 
         Returns
         -------
@@ -97,6 +128,7 @@ class UNFCCCApiReader:
         corresponding methods :py:meth:`UNFCCCApiReader.annex_one_reader.query` and
         :py:meth:`UNFCCCApiReader.non_annex_one_reader.query`.
         """
+        # select corresponding reader
         if party_code in self.annex_one_reader.parties["code"].values:
             reader = self.annex_one_reader
         elif party_code in self.non_annex_one_reader.parties["code"].values:
@@ -105,7 +137,12 @@ class UNFCCCApiReader:
             help = "try `UNFCCCApiReader().parties` for a list of valid codes"
             raise ValueError(f"Unknown party `{party_code}`, {help}!")
 
-        return reader.query(party_codes=[party_code], gases=gases, progress=progress)
+        return reader.query(
+            party_codes=[party_code],
+            gases=gases,
+            progress=progress,
+            normalize_gas_names=normalize_gas_names,
+        )
 
 
 class UNFCCCSingleCategoryApiReader:
@@ -258,6 +295,7 @@ class UNFCCCSingleCategoryApiReader:
         gases: typing.Union[None, typing.List[str]] = None,
         batch_size: int = 1000,
         progress: bool = False,
+        normalize_gas_names: bool = True,
     ) -> pd.DataFrame:
         """Query the UNFCCC for data.
 
@@ -280,13 +318,17 @@ class UNFCCCSingleCategoryApiReader:
             Default: query for all measures.
         gases : list of str, optional
             Limit the query to these gases. For possible values, see
-            :py:attr:`~UNFCCCApiReader.gases`. Default: query for all gases.
+            :py:attr:`~UNFCCCApiReader.gases`. Accepts subscripts ("N₂O")
+            as well as ASCII-strings ("N2O"). Default: query for all gases.
         batch_size : int, optional
             Number of variables to query in a single API query in the same batch to
             avoid internal server errors. Larger queries are split automatically.
             The default is 1000, which seems to work fine.
         progress : bool
             Display a progress bar. Requires the :py:mod:`tqdm` library. Default: false.
+        normalize_gas_names : bool, optional
+            If :obj:`True`, return gases as ASCII strings ("N2O").
+            Else, return native UNFCCC notation ("N₂O"). Default: true.
 
 
         Returns
@@ -301,6 +343,10 @@ class UNFCCCSingleCategoryApiReader:
         .. _UNFCCC documentation: https://unfccc.int/process-and-meetings/\
 transparency-and-reporting/greenhouse-gas-data/data-interface-help#eq-7
         """
+        # format gases to subscript notation
+        if gases is not None:
+            gases = [GAS_MAPPING.get(g, g) for g in gases]
+
         party_ids = []
         for code in party_codes:
             try:
@@ -342,7 +388,13 @@ transparency-and-reporting/greenhouse-gas-data/data-interface-help#eq-7
         if progress:
             pbar.close()
 
-        return self._parse_raw_answer(raw_response)
+        df = self._parse_raw_answer(raw_response)
+
+        if normalize_gas_names:
+            for c in ["unit", "gas"]:
+                df[c] = df[c].apply(lambda x: x.translate(MAKE_ASCII))
+
+        return df
 
     def _parse_raw_answer(self, raw: typing.List[dict]) -> pd.DataFrame:
         data = []